server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.5.4
Django==5.0.6
django-ipware==7.0.1
frozenlist==1.4.1
geoip2==4.8.0
idna==3.7
iniconfig==2.0.0
maxminddb==2.6.2
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-dotenv==1.0.1
python-ipware==3.0.0
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
